+++ spec/property.spec.ts
@@ -13,6 +13,6 @@ describe(`property tests`, () => {
         expect(<any>text instanceof Property).toBeTruthy(`the p function did not create an property class object`);
         expect(typeof text['type'] === 'function').toBeTruthy(`the property type is not a constructor`);
         expect(text['type'].name === 'String').toBeTruthy(`the property class did not assign a string type correctly`);
-        expect(text['default'] === "John Doe").toBeTruthy(``);
+        expect(text['default'] === "John Doe").toBeTruthy(`the property did not assign the default value correctly`);
     });
 });